add delete confirmation and loading ui

+++ components/transaction-item-remove-button.js
@@ -1,8 +1,32 @@
 import Button from './button';
 import { deleteTransaction } from '@/lib/actions';
+import { X, Loader } from 'lucide-react';
+import { useState } from 'react';
 
 export default function TransactionItemRemoveButton({ id }) {
-  return <Button onClick={async () => {
-    await deleteTransaction(id)
-  }}> x </Button>
+  const [loading, setLoading] = useState();
+  const [confirmed, setConfirmed] = useState();
+
+  const handleClick = async () => {
+    if (!confirmed) {
+      setConfirmed(true)
+      return
+    }
+    try {
+      setLoading(true)
+      await deleteTransaction(id)
+      //notify parent
+    } finally {
+      setLoading(false)
+    }
+  }
+
+  return (
+    <Button
+      size= 'xs' variant={!confirmed ? 'ghost' : 'danger'}
+      onClick={handleClick} aria-disabled={loading}>
+        {!loading && < X className='w-4 h-4' />}
+        {loading && < Loader className='w-4 h-4 animate-spin' />}
+    </Button>
+  )
 }
